+++ src/pages/ForgotPasswordPage/ForgotPasswordPage.jsx
@@ -1,0 +1,7 @@
+import ForgotPasswordForm from "../../components/ForgotPasswordForm/ForgotPasswordForm";
+
+const ForgotPasswordPage = () => {
+  return <ForgotPasswordForm />;
+};
+
+export default ForgotPasswordPage;
