+++ src/pages/ResetPasswordPage/ResetPasswordPage.jsx
@@ -1,0 +1,7 @@
+import ResetPasswordForm from "../../components/ResetPasswordForm/ResetPasswordForm";
+
+const ResetPasswordPage = () => {
+  return <ResetPasswordForm />;
+};
+
+export default ResetPasswordPage;
